domain entity added (#31)
* domain entity added
namespace corrected same as domain project name

* move dispencer status to shared assembly

* pr comments resolved. 2 issues still require discussion

* moved gethashcode and equals to entity base
moved usage calculation to usage itself

* renamed all occurrences of Dispenser +fix other comment

* pr comments fixed

+++ BeerDispancer/BeerDispencer.Application/Implementation/Response/UsageResponse.cs
@@ -5,7 +5,7 @@ namespace BeerDispancer.Application.Implementation.Response
 {
     public class UsageResponse
     {
-        public double Amount { get; set; } 
+        public decimal Amount { get; set; } 
         public UsageEntry[] Usages { get; set; }
     }
 
@@ -13,8 +13,8 @@ namespace BeerDispancer.Application.Implementation.Response
     {
         public DateTime? OpenedAt { get; set; }
         public DateTime? ClosedAt { get; set; }
-        public double? FlowVolume { get; set; }
-        public double? TotalSpent { get; set; }
+        public decimal? FlowVolume { get; set; }
+        public decimal? TotalSpent { get; set; }
 
     }
 }
